Rename folder to `particles`.
