+++ particles/resources/program_fragment_shader.glsl
